refactor : routes폴더 생성 및 전용 컴포넌트 분리
